Revert "Revert "fix(router): Ensure canMatch guards run on wildcard routes (#53239)" (#53339)"
This reverts commit ac8eb5b943d34b8ab68a51b3d46cedafa037d03d.
